+++ src/scenes/scrolling_name.cpp
@@ -16,7 +16,7 @@ void ScrollingNamesScene::Update(Window &wind) {
 	textPosition -= scrollSpeed * wind.deltaTime;
 
 	// jump back to the other side of the screen, but offscreen so we can scroll in
-	int resetPoint = -((int) strlen(name) * FontSize() * textScale);
+	int resetPoint = -(int)(strlen(name) * FontSize() * textScale);
 	if (textPosition <= resetPoint) {
 		textPosition = wind.w;
 	}
@@ -25,7 +25,7 @@ void ScrollingNamesScene::Update(Window &wind) {
 void ScrollingNamesScene::Render(Window &wind) {
 	wind.Clear(ColorFromRGB(0, 0, 0));
 	wind.DrawString(
-		textPosition,
+		(int) textPosition,
 		wind.h / 2 - FontSize() * textScale / 2,
 		textScale,
 		name,
